// packages/control-property-editor/src/api.ts
export type PropertyValue = string | boolean | number;

export interface PropertyChange<T extends PropertyValue = PropertyValue> {
    controlId: string;
    controlName: string;
    propertyName: string;
    value: T;
}
export interface PropertyChanged<T extends PropertyValue = PropertyValue> {
    controlId: string;
    propertyName: string;
    newValue: T;
}

export interface PropertyChangeFailed {
    controlId: string;
    propertyName: string;
    errorMessage: string;
}

export const BOOLEAN_VALUE_TYPE = 'boolean';
export const INTEGER_VALUE_TYPE = 'integer';
export const FLOAT_VALUE_TYPE = 'float';
export const STRING_VALUE_TYPE = 'string';
export const INPUT_EDITOR_TYPE = 'input';
export const DROPDOWN_EDITOR_TYPE = 'dropdown';
export const CHECKBOX_EDITOR_TYPE = 'checkbox';

interface ControlPropertyBase<T, V, E> {
    type: T;
    editor: E;
    name: string;
    readableName: string;
    value: V | string; // string: expression binding
    isEnabled: boolean;
    isIcon?: boolean;
    label?: string;
    errorMessage?: string;
    documentation?: {
        defaultValue: string;
        description: string;
        propertyName: string;
        type?: string;
        propertyType?: string;
    };
}
// ...
export interface StringControlPropertyWithOptions
    extends ControlPropertyBase<typeof STRING_VALUE_TYPE, string, typeof DROPDOWN_EDITOR_TYPE> {
    options: { key: string; text: string }[];
}